Banyak Perubahan

+++ admin/dataproduk.php
@@ -1,5 +1,5 @@
 <div class="d-sm-flex align-items-center justify-content-between mb-4">
-  <h1 class="h3 mb-0 text-gray-800"><i class="fas fa-fw fa-tablets"></i> Data Produk</h1>
+  <h1 class="h3 mb-0 text-gray-800"><i class="fas fa-boxes"></i> Data Produk</h1>
 
   <a href="?page=inputproduk" class="btn btn-success"> <i class="fa fa-plus"></i> Tambah Data </a>
 </div>
@@ -20,6 +20,7 @@
             <th>Nama Produk</th>
             <th>Tanggal Exp</th>
             <th>Kategori</th>
+            <th>Satuan</th>
             <th>Stok</th>
             <th>Harga Beli</th>
             <th>Harga Jual</th>
@@ -29,7 +30,7 @@
         <tbody>
           <?php
           $no = 0;
-          $query = "SELECT a.kd, a.nama, b.tgl_exp, c.nama as kategori, b.harga_beli, a.harga_jual, b.stok
+          $query = "SELECT a.kd, a.nama, b.tgl_exp, c.nama as kategori, a.satuan, b.harga_beli, a.harga_jual, b.stok
 										FROM produk a 
 										LEFT JOIN detail_produk b ON a.kd=b.kd_produk
 										LEFT JOIN kategori c ON a.kd_kategori=c.kd
@@ -37,27 +38,23 @@
           $result  = mysqli_query($konek, $query);
           while ($data = mysqli_fetch_assoc($result)) {
           ?>
-          <tr class="text-center">
-            <td><?php echo ++$no; ?></td>
-            <td><?php echo $data['kd']; ?></td>
-            <td align="left"><?php echo $data['nama']; ?></td>
-            <td><?php echo $data['tgl_exp']; ?></td>
-            <td align="left"><?php echo $data['kategori']; ?></td>
-            <td><?php echo $data['stok']; ?></td>
-            <td align="right"><?php echo "Rp " . number_format($data['harga_beli'], 0, ',', '.'); ?></td>
-            <td align="right"><?php echo "Rp " . number_format($data['harga_jual'], 0, ',', '.'); ?></td>
-            <td>
-              <div class="btn-group" role="group">
-                <a data-toggle="tooltip" data-placement="bottom" title="Edit Data"
-                  href="?page=editproduk&id=<?php echo $data['kd']; ?>" class="btn btn-warning btn-sm"><i
-                    class="fa fa-edit"></i></a>
-                <a data-toggle="tooltip" data-placement="bottom" title="Hapus Data"
-                  href="hapusproduk.php?id=<?php echo $data['kd']; ?>"
-                  onclick="return confirm ('Apakah anda yakin untuk meghapus data ini')"
-                  class="btn btn-danger btn-sm"><i class="fa fa-trash"></i></a>
-              </div>
-            </td>
-          </tr>
+            <tr class="text-center">
+              <td><?php echo ++$no; ?></td>
+              <td><?php echo $data['kd']; ?></td>
+              <td align="left"><?php echo $data['nama']; ?></td>
+              <td><?php echo $data['tgl_exp']; ?></td>
+              <td align="left"><?php echo $data['kategori']; ?></td>
+              <td><?php echo $data['satuan']; ?></td>
+              <td><?php echo $data['stok']; ?></td>
+              <td align="right"><?php echo "Rp " . number_format($data['harga_beli'], 0, ',', '.'); ?></td>
+              <td align="right"><?php echo "Rp " . number_format($data['harga_jual'], 0, ',', '.'); ?></td>
+              <td>
+                <div class="btn-group" role="group">
+                  <a data-toggle="tooltip" data-placement="bottom" title="Edit Data" href="?page=editproduk&id=<?php echo $data['kd']; ?>" class="btn btn-warning btn-sm"><i class="fa fa-edit"></i></a>
+                  <a data-toggle="tooltip" data-placement="bottom" title="Hapus Data" href="hapusproduk.php?id=<?php echo $data['kd']; ?>" onclick="return confirm ('Apakah anda yakin untuk meghapus data ini')" class="btn btn-danger btn-sm"><i class="fa fa-trash"></i></a>
+                </div>
+              </td>
+            </tr>
           <?php
           }
           ?>
